+++ src/main/java/com/sarmich/timetable/model/request/TeacherUpdateRequest.java
@@ -5,10 +5,10 @@ import jakarta.validation.constraints.NotNull;
 
 public record TeacherUpdateRequest(
         @NotNull(message = "id required")
-        Long id,
+        Integer id,
         String firstName,
         String lastName,
-        Long subjectId,
+        Integer subjectId,
         LessonTime lessonTimes
 ) {
 }
